requirements.txt
# APT Detection System - Requirements
# Core ML/DL Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1
xgboost==1.7.6

# Data Processing
scipy==1.11.1
imbalanced-learn==0.11.0

# Network Analysis
scapy==2.5.0
pyshark==0.6

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Dashboard
streamlit==1.25.0
dash==2.11.1
dash-bootstrap-components==1.4.2

# Threat Intelligence
stix2==3.0.1
taxii2-client==2.3.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
tqdm==4.65.0
joblib==1.3.1
requests==2.31.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Documentation
jupyter==1.0.0
notebook==7.0.0
ipywidgets==8.0.7

# Model Interpretation
shap==0.42.1
lime==0.2.0.1

# Performance Monitoring
mlflow==2.5.0
